LPD-29771 Apply rename to tests
