+++ modules/test/playwright/tests/frontend-data-set-admin-web/tests/data-set-fragment/pages/FDSFragmentPage.ts
@@ -101,7 +101,7 @@ export class FDSFragmentPage {
 
 		await this.page
 			.frameLocator('iframe[title="Select"]')
-			.locator('.fds-view-item-selector')
+			.locator('.fds-admin-item-selector')
 			.waitFor({state: 'visible'});
 
 		await this.page
